Added user authentication to endpoints in Rental controller class

+++ backend/src/main/java/no/ntnu/project/group4/webapp/controllers/RentalController.java
@@ -1,6 +1,7 @@
 package no.ntnu.project.group4.webapp.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -12,6 +13,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import no.ntnu.project.group4.webapp.models.Rental;
+import no.ntnu.project.group4.webapp.models.User;
+import no.ntnu.project.group4.webapp.services.AccessUserService;
 import no.ntnu.project.group4.webapp.services.RentalService;
 
 import java.util.Optional;
@@ -22,6 +25,8 @@ import java.util.Optional;
 public class RentalController {
   @Autowired
   private RentalService rentalService;
+  @Autowired
+  private AccessUserService userService;
 
   @GetMapping("/get")
   public Iterable<Rental> getAll() {
@@ -29,10 +34,24 @@ public class RentalController {
   }
 
   @GetMapping("/get/{id}")
-  public ResponseEntity<Rental> getOne(@PathVariable Long id) {
-    ResponseEntity<Rental> response;
+  public ResponseEntity<?> getOne(@PathVariable Long id) {
+    ResponseEntity<?> response;
+    User sessionUser = userService.getSessionUser();
     Optional<Rental> rental = rentalService.getOne(id);
-    response = rental.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+    if (rental.isPresent()) {
+      Rental foundRental = rental.get();
+      if (sessionUser != null && sessionUser.getEmail().equals(foundRental.getUser().getEmail())) {
+        response = new ResponseEntity<>(foundRental, HttpStatus.OK);
+      } else if (sessionUser == null) {
+        response = new ResponseEntity<>("Rental data accessible only to authenticated users",
+                                        HttpStatus.UNAUTHORIZED);
+      } else {
+        response = new ResponseEntity<>("Rental data for other users not accessible",
+                                        HttpStatus.FORBIDDEN);
+      }
+    } else {
+      response = new ResponseEntity<>("", HttpStatus.NOT_FOUND);
+    }
     return response;
   }
 
@@ -49,12 +68,25 @@ public class RentalController {
   }
 
   @DeleteMapping("/del/{id}")
-  public ResponseEntity<Void> deleteRental(@PathVariable Long id) {
-    boolean wasDeleted = rentalService.delete(id);
-    if (wasDeleted) {
-      return ResponseEntity.ok().build();
+  public ResponseEntity<String> deleteRental(@PathVariable Long id) {
+    ResponseEntity<String> response;
+    User sessionUser = userService.getSessionUser();
+    Optional<Rental> rental = rentalService.getOne(id);
+    if (rental.isPresent()) {
+      if (sessionUser != null &&
+          sessionUser.getEmail().equals(rental.get().getUser().getEmail())) {
+        rentalService.delete(id);
+        response = new ResponseEntity<>("", HttpStatus.OK);
+      } else if (sessionUser == null) {
+        response = new ResponseEntity<>("Rental data accessible only to authenticated users",
+                                        HttpStatus.UNAUTHORIZED);
+      } else {
+        response = new ResponseEntity<>("Rental data for other users not accessible",
+                                        HttpStatus.FORBIDDEN);
+      }
     } else {
-      return ResponseEntity.notFound().build();
+      response = new ResponseEntity<>("", HttpStatus.NOT_FOUND);
     }
+    return response;
   }
 }
